+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.6 [????.??.??]
+* Allow building with `template-haskell-2.18.0.0` (GHC 9.2).
 * `deriveVia` now imposes the requirement that the type variables mentioned in
   the second argument to `Via` must be a subset of the type variables mentioned
   in the instance context and the first argument to `Via`. In effect, this